# src/main/resources/swagger/swagger-input.yml
swagger: "2.0"
info:
  description: "PPM Info Adapter"
  version: "1.0.0"
  title: "PPM Info Adapter"
tags:
- name: "patients"
  description: "Everything about Personal Info"
schemes:
- "https"
- "http"
paths:
  /v1/patients/info:
    post:
      tags:
        - "patients"
      summary: "Upsert patient personal info"
      description: "Upsert patient personal info"
      operationId: "upsert"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "info"
          description: "Patient personal info"
          required: true
          schema:
            $ref: "#/definitions/PersonalInfo"
      responses:
        200:
          description: "OK"
          schema:
            type: "string"
        400:
          description: "Invalid input"
        500:
          description: "Internal server error"
  /v1/patients/info/bulk:
    post:
      tags:
        - "patients"
      summary: "find multi patient personal info"
      description: "find multi patient personal info"
      operationId: "readMultiPatientInfoByPatientIds"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "ids"
          description: "ids of patient"
          required: true
          schema:
            type: "array"
            items:
              type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PersonalInfo"
        400:
          description: "Invalid input"
        500:
          description: "Internal server error"

  /v1/patients/info/names:
    get:
      tags:
      - "patients"
      summary: "search names of patients"
      description: "search names of patients"
      operationId: "searchPatientsByName"
      produces: "application/json"
      parameters:
      - in: "query"
        name: "patientName"
        description: "name of a patient, note that name of patients can be the same"
        required: "false"
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PersonalInfo"

        400:
          description: "Invalid input"
        500:
          description: "Internal server error"
  /v1/patients/{patientId}/info:
    get:
      tags:
      - "patients"
      summary: "Get patient personal info by id"
      description: "Get patient personal info by id"
      operationId: "readPatientInfoById"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "patientId"
        description: "Id of patient to get"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/PersonalInfo"
        400:
          description: "Invalid input"
        404:
          description: "PatientEntity not found"
        500:
          description: "Internal server error"
definitions:
  PersonalInfo:
    type: "object"
    properties:
      patientId:
        type: "string"
      fullname:
        type: "string"
        example: "Nguyen Minh Duc"
      address:
        type: "string"
        example: "3/115 Binh Duong"
      dob:
        type: "string"
        format: "date"
        example: "1995-09-18"
      pob:
        type: "string"
        example: "Binh Duong"
      sex:
        type: "string"
        example: "M"